+++ scylla/src/transport/mod.rs
@@ -21,7 +21,6 @@ pub mod topology;
 pub use crate::frame::{Authenticator, Compression};
 pub use connection::SelfIdentity;
 pub use execution_profile::ExecutionProfile;
-pub use scylla_cql::errors;
 pub use scylla_cql::frame::request::query::{PagingState, PagingStateResponse};
 
 #[cfg(test)]
