+++ scylla/src/transport/retry_policy.rs
@@ -2,9 +2,8 @@
 //! To decide when to retry a query the `Session` can use any object which implements
 //! the `RetryPolicy` trait
 
-use crate::errors::QueryError;
+use crate::errors::{DbError, QueryError, WriteType};
 use crate::frame::types::Consistency;
-use crate::transport::errors::{DbError, WriteType};
 
 /// Information about a failed query
 pub struct QueryInfo<'a> {
@@ -221,9 +220,9 @@ impl RetrySession for DefaultRetrySession {
 mod tests {
     use super::{DefaultRetryPolicy, QueryInfo, RetryDecision, RetryPolicy};
     use crate::errors::{BadQuery, QueryError};
+    use crate::errors::{DbError, WriteType};
     use crate::statement::Consistency;
     use crate::test_utils::setup_tracing;
-    use crate::transport::errors::{DbError, WriteType};
     use bytes::Bytes;
     use std::io::ErrorKind;
     use std::sync::Arc;
